4.17-4.23 done

+++ controllers/blogs.js
@@ -1,26 +1,46 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const {userExtractor} = require('../utils/middleware')
+
+/* const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+} */
 
 blogsRouter.get('/', async (request, response, next) => {
   try {
-    const blogs = await Blog.find({}).populate('user',{blogs:0})
+    const blogs = await Blog.find({}).populate('user', { blogs: 0 })
     response.json(blogs)
   } catch (exception) {
     next(exception)
   }
-    
-})
-  
-blogsRouter.post('/', async (request, response,next) => {
-  const user = await User.findById(request.body.userId)
-  const blog = new Blog({...request.body,user: user._id})
 
-  blog.likes = blog?.likes ?? 0
-  if (!(blog?.title && blog?.url)) response.status(400).json('title or url missing')
-  try {
+})
+//register a middleware userExtractor only for a specific operation
+blogsRouter.post('/',userExtractor, async (request, response, next) => {
+  //const user = await User.findById(request.body.userId)
+  //const token = getTokenFrom(request)
+  //const decodedToken = jwt.verify(token, process.env.SECRET)
+  try { 
+    
+    /* const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id) */
+    const user = request.user
+    const blog = new Blog({ ...request.body, user: user._id })
+
+    blog.likes = blog?.likes ?? 0
+    if (!(blog?.title && blog?.url)) response.status(400).json('title or url missing')
+
     const savedBlog = await blog.save()
-    user.blogs = [...user.blogs, savedBlog._id]  
+    user.blogs = [...user.blogs, savedBlog._id]
     await user.save()
     response.status(201).json(savedBlog)
   } catch (exception) {
@@ -41,10 +61,26 @@ blogsRouter.get('/:id', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    
+/*     if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    } */
+    const user = request.user
+    if (blog.user.toString() === user.id){
+      await Blog.findByIdAndRemove(request.params.id)
+      await User.findOneAndUpdate(
+        { username: decodedToken.username },
+        { $pull: { blogs: { $in:[request.params.id] }} },
+        { new: true }
+      )
+      response.status(204).end()
+    } else {
+      response.status(401).end()
+    }
   } catch (exception) {
     next(exception)
   }
@@ -65,7 +101,7 @@ blogsRouter.put('/:id', async (request, response, next) => {
   } catch (exception) {
     next(exception)
   }
- 
+
 })
 
 module.exports = blogsRouter
